Add convenience methods to get double value from command arguments and provide support for decimal values in Boss commands

+++ src/main/java/org/mineacademy/boss/command/EggDropCommand.java
@@ -39,9 +39,9 @@ final class EggDropCommand extends BossEggAbstractCommand {
 	protected void onCommand() {
 		final World world = this.findWorld(this.args[0]);
 
-		final int x = this.parseCoordinate(1);
-		final int y = this.parseCoordinate(2);
-		final int z = this.parseCoordinate(3);
+		final double x = this.parseCoordinate(1);
+		final double y = this.parseCoordinate(2);
+		final double z = this.parseCoordinate(3);
 
 		final Location location = new Location(world, x, y, z);
